Added a lot
Technically this works, but I would like to find a better way to get data into the database without so many things depending on eachother. Look into caching files. subproccess or thinkg of some other ideas to return data in a loop or get data to the parser without breaking the loop.